backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pandas>=2.2.0
python-multipart>=0.0.9
pydantic>=2.6.1
plotly>=5.18.0
groq>=0.4.1
python-dotenv>=1.0.1
openpyxl>=3.1.2
supabase>=2.3.5
psycopg2-binary>=2.9.9
httpx>=0.25.0,<0.28.0
fpdf2>=2.7.8
kaleido==0.2.1